// src/shared/api/model/responsePosters.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * Anime ON API
 * OpenAPI spec version: ##VERSION## DEV
 */
import type { ChildAnimePoster } from './childAnimePoster';

export interface ResponsePosters {
  anime?: ChildAnimePoster;
  description?: string;
  /** @nullable */
  readonly image?: string | null;
}